No longer use architecture specific integer types when not necessary (#354)
* Use u64 for transfer_batch_size_in_bytes

* Use u64 for memory_in_bytes fields in configuration

* Use i64 for MemoryPool

* Use u64 for transfer_time_in_seconds

* Use u8 for the number of threads for ingestion, compression, and writing

* Remove unused test const

* Run Rustfmt

* Add missing semicolon

* Fix type issue after merge

* Use constant for all expect messages related to converting size in bytes

* Run Rustfmt

+++ crates/modelardb_server/src/storage/mod.rs
@@ -214,7 +214,7 @@ impl StorageEngine {
     /// Start `num_threads` threads with `name` that executes `function` and whose [`JoinHandle`] is
     /// added to `join_handles.
     fn start_threads<F>(
-        num_threads: usize,
+        num_threads: u8,
         name: &str,
         function: F,
         join_handles: &mut Vec<JoinHandle<()>>,
@@ -261,7 +261,7 @@ impl StorageEngine {
     ) -> Result<()> {
         // TODO: write to a WAL and use it to ensure termination never duplicates or loses data.
         self.memory_pool
-            .wait_for_ingested_memory(multivariate_data_points.get_array_memory_size());
+            .wait_for_ingested_memory(multivariate_data_points.get_array_memory_size() as u64);
 
         self.channels
             .ingested_data_sender
@@ -313,7 +313,7 @@ impl StorageEngine {
     }
 
     /// Change the amount of memory for multivariate data in bytes according to `value_change`.
-    pub(super) async fn adjust_multivariate_remaining_memory_in_bytes(&self, value_change: isize) {
+    pub(super) async fn adjust_multivariate_remaining_memory_in_bytes(&self, value_change: i64) {
         self.memory_pool.adjust_ingested_memory(value_change)
     }
 
@@ -322,7 +322,7 @@ impl StorageEngine {
     /// spilled.
     pub(super) async fn adjust_uncompressed_remaining_memory_in_bytes(
         &self,
-        value_change: isize,
+        value_change: i64,
     ) -> Result<()> {
         self.uncompressed_data_manager
             .adjust_uncompressed_remaining_memory_in_bytes(value_change)
@@ -333,7 +333,7 @@ impl StorageEngine {
     /// value is changed successfully return [`Ok`], otherwise return [`ModelarDbServerError`].
     pub(super) async fn adjust_compressed_remaining_memory_in_bytes(
         &self,
-        value_change: isize,
+        value_change: i64,
     ) -> Result<()> {
         self.compressed_data_manager
             .adjust_compressed_remaining_memory_in_bytes(value_change)
@@ -352,7 +352,7 @@ impl StorageEngine {
 
     /// Remove the table with `table_name` from the tables that are marked as dropped and clear the
     /// size of the table in the data transfer component. Return the number of bytes that were cleared.
-    pub(super) async fn clear_table(&self, table_name: &str) -> usize {
+    pub(super) async fn clear_table(&self, table_name: &str) -> u64 {
         if let Some(ref mut data_transfer) =
             *self.compressed_data_manager.data_transfer.write().await
         {
@@ -367,7 +367,7 @@ impl StorageEngine {
     /// [`ModelarDbServerError`].
     pub(super) async fn set_transfer_batch_size_in_bytes(
         &self,
-        new_value: Option<usize>,
+        new_value: Option<u64>,
     ) -> Result<()> {
         if let Some(ref mut data_transfer) =
             *self.compressed_data_manager.data_transfer.write().await
@@ -386,7 +386,7 @@ impl StorageEngine {
     /// transfer component does not exist, return [`ModelarDbServerError`].
     pub(super) async fn set_transfer_time_in_seconds(
         &mut self,
-        new_value: Option<usize>,
+        new_value: Option<u64>,
     ) -> Result<()> {
         if let Some(ref mut data_transfer) =
             *self.compressed_data_manager.data_transfer.write().await
